chore: move builder from grpc to uopool, refactor examples, storage examples
