+++ crates/grpc/Cargo.toml
@@ -17,7 +17,6 @@ dashmap = "5.4.0"
 ethers = { workspace = true }
 expanded-pathbuf = { workspace = true }
 eyre = { workspace = true }
-futures-util = { workspace = true }
 parking_lot = { workspace = true }
 prost = "0.11"
 serde_json = { workspace = true }
