+++ crates/grpc/src/lib.rs
@@ -1,6 +1,5 @@
 #![allow(dead_code)]
 
-mod builder;
 mod bundler;
 mod proto;
 mod uopool;
