--- examples/Cargo.toml
@@ -1,44 +1,0 @@
-[package]
-name = "silius-examples"
-version = { workspace = true }
-authors = { workspace = true }
-edition = { workspace = true }
-license = { workspace = true }
-repository = { workspace = true }
-rust-version = { workspace = true }
-readme = "README.md"
-description = """
-AA (ERC-4337) bundler examples
-"""
-
-[dependencies]
-alloy-primitives = "0.3"
-alloy-sol-types = "0.3"
-ethers = { workspace = true }
-eyre = { workspace = true }
-reqwest = { version = "0.11.4", features = ["json"] }
-serde = "1"
-serde_json = "1"
-silius-contracts = { path = "../crates/contracts" }
-silius-primitives = { path = "../crates/primitives" }
-silius-tests = { path = "../tests" }
-tokio = { workspace = true }
-
-[package.metadata.cargo-udeps.ignore]
-normal = ["eyre", "silius-contracts", "silius-primitives", "silius-tests"]
-
-[[example]]
-name = "user-operation"
-path = "user_operation.rs"
-
-[[example]]
-name = "simple-account-create"
-path = "simple_account/create.rs"
-
-[[example]]
-name = "simple-account-deposit"
-path = "simple_account/deposit.rs"
-
-[[example]]
-name = "simple-account-transfer"
-path = "simple_account/transfer.rs"
