+++ examples/simple-account/Cargo.toml
@@ -1,0 +1,39 @@
+[package]
+name = "examples-simple-account"
+version = { workspace = true }
+authors = { workspace = true }
+edition = { workspace = true }
+license = { workspace = true }
+repository = { workspace = true }
+rust-version = { workspace = true }
+description = """
+AA (ERC-4337) bundler examples - simple account
+"""
+
+[dependencies]
+alloy-primitives = "0.3"
+alloy-sol-types = "0.3"
+ethers = { workspace = true }
+eyre = { workspace = true }
+reqwest = { version = "0.11.4", features = ["json"] }
+serde = "1"
+serde_json = "1"
+silius-contracts = { path = "../../crates/contracts" }
+silius-primitives = { path = "../../crates/primitives" }
+silius-tests = { path = "../../tests" }
+tokio = { workspace = true }
+
+[package.metadata.cargo-udeps.ignore]
+normal = ["silius-contracts", "silius-primitives", "silius-tests"]
+
+[[example]]
+name = "create"
+path = "examples/create.rs"
+
+[[example]]
+name = "deposit"
+path = "examples/deposit.rs"
+
+[[example]]
+name = "transfer"
+path = "examples/transfer.rs"
