+++ examples/simple-account/src/lib.rs
