+++ examples/storage/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "examples-storage"
+version = { workspace = true }
+authors = { workspace = true }
+edition = { workspace = true }
+license = { workspace = true }
+repository = { workspace = true }
+rust-version = { workspace = true }
+description = """
+AA (ERC-4337) bundler examples - storage
+"""
+
+[dependencies]
+ethers = { workspace = true }
+eyre = { workspace = true }
+silius-primitives = { path = "../../crates/primitives" }
+silius-uopool = { path = "../../crates/uopool" }
+tempdir = "0.3.7"
+tokio = { workspace = true }
+
+[[example]]
+name = "memory"
+path = "examples/memory.rs"
+
+[[example]]
+name = "database"
+path = "examples/database.rs"
+
